Relationship tables finished + auth fixed + excpetions fixed

+++ EntitiesAPI/spring-server/src/main/java/ch/heigvd/amt/project2/repositories/IMovieRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IMovieRepository extends CrudRepository<MovieEntity, Long> {
+public interface IMovieRepository extends CrudRepository<MovieEntity, Integer> {
 
     Page<MovieEntity> findAll(Pageable pageable);
 }
